physics in seperate file
